bcp de choses de fait mais j'ai codé sur lmain sry

+++ src/sample/Rue.java
@@ -8,7 +8,7 @@ import javafx.scene.shape.Line;
 import java.util.ArrayList;
 
 public class Rue {
-    double longueur;
+    long longueur;
     String nom;
     Line trait = new Line();
 
@@ -21,8 +21,8 @@ public class Rue {
 
     //ArrayList<Maison> maisons = new ArrayList<>();
 
-    public double getLongueur() {return longueur;}
-    public void setLongueur(double longueur) {this.longueur = longueur;}
+    public long getLongueur() {return longueur;}
+    public void setLongueur(long longueur) {this.longueur = longueur;}
     public String getNom() {return nom;}
     public void setNom(String nom) {this.nom = nom;}
     //public ArrayList<Maison> getMaisons() {return maisons;}
